pyproject.toml
[tool.poetry]
name = "ecommerce-django-api"
version = "0.1.0"
description = ""
authors = ["nuriddin <sobitovnuriddin7655@gmail.com>"]
readme = "README.md"
packages = [{include = "ecommerce_django_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
django-mptt = "^0.14.0"
pillow = "^9.5.0"
djangorestframework-simplejwt = "^5.2.2"
boto3 = "^1.26.158"
django-storages = "^1.13.2"
minio = "^7.1.15"
django-filter = "^23.2"
coreapi = "^2.3.3"
pyyaml = "^6.0"
django-elasticsearch-dsl-drf = "^0.22.5"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
